Normalized damage constants to damage done to the chest. This means all modifiers were divided by 1.85 and all base damages were multiplied by 1.85. I also removed the stupid 20% bayonet damage reduction hack and rolled that into their base damages (multiplying the damages by 0.8). Finally, I noticed some weapons use identical base damage values, so I had them equal each other. These weapons include: - Brown Bess and Pennsylvania - Charleville and Revolutionnaire
/Tjoppen

+++ game/server/bg2/weapon_bg2base.cpp
@@ -419,7 +419,7 @@ void CBaseBG2Weapon::Hit( trace_t &traceHit, int iAttack )
 		pPlayer->EyeVectors( &hitDirection, NULL, NULL );
 		VectorNormalize( hitDirection );
 
-		float	damage		= GetDamage( iAttack );	
+		int damage		= GetDamage( iAttack );	
 
 		//BG2 - Tjoppen - apply no force
 		CTakeDamageInfo info( GetOwner(), GetOwner(), damage, DMG_CLUB | DMG_PREVENT_PHYSICS_FORCE | DMG_NEVERGIB );
